+++ src/app/_internal/translate.module.ts
@@ -1,0 +1,25 @@
+import { NgModule } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+
+// Create a factory for the translate loader
+export function createTranslateLoader(http: HttpClient) {
+    return new TranslateHttpLoader(http, '../assets/i18n/', '.json');
+}
+
+@NgModule({
+    imports: [
+        TranslateModule.forChild({
+            loader: {
+                provide: TranslateLoader,
+                useFactory: (createTranslateLoader),
+                deps: [HttpClient]
+            },
+            isolate: false
+        })
+    ],
+    exports: [TranslateModule]
+})
+export class TranslateSharedLazyModule {
+}