+++ src/app/pages/i18n/i18n.module.ts
@@ -12,18 +12,21 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 /***** Basic app components *****/
 import { RoutesModule } from './routes';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { UtilsModule } from '../../_internal/utils/utils.module';
 
 /***** Import WUF *****/
 import { WufLayoutModule } from '@anviltech/wuf-ang-layout';
 
 /***** 3rd party imports *****/
 import { NgxMdModule } from 'ngx-md';
+import { CustomMaterialModule } from '../../_internal/material.module';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { TranslateSharedLazyModule } from '../../_internal/translate.module';
 
 /***** Import page modules *****/
-import { DevAngularAngularComponentsIntroComponent } from './dev-angular-components/dev-angular-components.component';
-import { DevWebAngularComponentsIntroComponent } from './dev-web-components/dev-web-components.component';
-import { DevelopmentComponent } from './development/development.component';
+import { I18nIntroComponent } from './i18n-intro/i18n-intro.component';
+import { I18nAngularComponent } from './i18n-angular/i18n-ang.component';
+import { I18nNgxTranslateComponent } from './i18n-ngx-translate/i18n-ngx-translate.component';
 
 
 @NgModule({
@@ -40,8 +43,13 @@ import { DevelopmentComponent } from './development/development.component';
         // WUF
         WufLayoutModule,
 
+        // Utils
+        UtilsModule,
+
         // 3rd Party Imports
         NgxMdModule, // Markdown
+        CustomMaterialModule,
+        TranslateModule,
         TranslateSharedLazyModule,
 
         // Routes (Keep as last module loaded)
@@ -50,10 +58,13 @@ import { DevelopmentComponent } from './development/development.component';
     ],
     declarations: [
         // Page imports
-        DevAngularAngularComponentsIntroComponent,
-        DevWebAngularComponentsIntroComponent,
-        DevelopmentComponent
+        I18nIntroComponent,
+        I18nAngularComponent,
+        I18nNgxTranslateComponent
+    ],
+    providers: [
+        TranslateService
     ]
 })
-export class DevelopmentModule {
+export class I18nModule {
 }