+++ projects/anviltech/wuf-ang-layout/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [2.0.0-rc.4] - 2019-03-21
+### Changed
+- Upgrading peer dependencies
+
 ## [2.0.0-rc.3] - 2019-03-13
 ### Changed
 - Nothing.  Just a test publish.