+++ projects/anviltech/wuf-ang-navigation/CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [2.0.0-rc.6] - 2019-03-21
+### Fixed
+- Add missing TranslateService to providers list.
+
+## [2.0.0-rc.5] - 2019-03-21
+### Fixed
+- upgrading dependencies
+
+## [2.0.0-rc.4] - 2019-03-21
+### Fixed
+- Adding missing WufSidebarService to providers list.
+
+## [2.0.0-rc.3] - 2019-03-21
+### Added
+- Adding support for ngx-translate.
+
 ## [2.0.0-rc.2] - 2019-03-13
 ### Fixed
 - Fix NavResponseData interface